+++ website/app/data/project.ts
@@ -1,5 +1,6 @@
 export const repository = 'https://github.com/ExtraBinoss/argui'
 export const discord = 'https://discord.gg/xY9CWSc65'
+export const currentVersion = '0.3.2'
 export const sourceUrl = (path: string) => `${repository}/blob/main/${path}`
 export const composeExample = `use argui::{
     ui::Element,